refactor(saml20): moves XML signing code into ./xml/sign
